+++ index.mjs
@@ -31,7 +31,9 @@ export default class Smarty {
     this.reset()
     this.__CACHE__ = Object.create(null) // 渲染缓存
     // 消除缓存目录
-    fs.rm(cacheDir)
+    if (fs.isdir(cacheDir)) {
+      fs.rm(cacheDir)
+    }
   }
 
   reset() {
